separate views for address and demands edition in admin role
